Customer ID clear is set to 0
On entering a new customer, back-end will fail to fetch any details. So, if any previously selected customer ID is present, that has to be cleared. On clearing, customer id field ends up with no value which throws error on submitting. So, its updated to 0 instead of clearing it out